IDE dependencies filter - filter by import list (#4412)
- Given a list of import lines, typically the import section of a module
  being edited, filter to the declarations brought into scope by those
  imports plus specified qualifier.
