+++ CHANGELOG.d/feature_ide-dependency-filter.md
@@ -1,0 +1,6 @@
+```markdown
+* Add `purs ide` dependency/imports filter (#4412 by @nwolverson)
+
+  This allows IDE tooling to filter type searches according to the imports of a given module,
+  restricting to identifiers in scope.
+```
