+++ purescript.cabal
@@ -286,7 +286,9 @@ library
     Language.PureScript.Ide.Externs
     Language.PureScript.Ide.Filter
     Language.PureScript.Ide.Filter.Declaration
+    Language.PureScript.Ide.Filter.Imports
     Language.PureScript.Ide.Imports
+    Language.PureScript.Ide.Imports.Actions
     Language.PureScript.Ide.Logging
     Language.PureScript.Ide.Matcher
     Language.PureScript.Ide.Prim
