+++ tests/Language/PureScript/Ide/FilterSpec.hs
@@ -6,13 +6,14 @@ import qualified Data.Set as Set
 import           Language.PureScript.Ide.Filter
 import           Language.PureScript.Ide.Filter.Declaration as D
 import           Language.PureScript.Ide.Types
+import           Language.PureScript.Ide.Imports
 import           Language.PureScript.Ide.Test as T
 import qualified Language.PureScript as P
 import           Test.Hspec
 
 type Module = (P.ModuleName, [IdeDeclarationAnn])
 
-moduleA, moduleB, moduleC, moduleD, moduleE, moduleF, moduleG, moduleH, moduleI :: Module
+moduleA, moduleB, moduleC, moduleD, moduleE, moduleF, moduleG, moduleH, moduleI, moduleDCtors :: Module
 moduleA = (P.moduleNameFromString "Module.A", [T.ideValue "function1" Nothing])
 moduleB = (P.moduleNameFromString "Module.B", [T.ideValue "data1" Nothing])
 moduleC = (P.moduleNameFromString "Module.C", [T.ideType "List" Nothing []])
@@ -22,9 +23,13 @@ moduleF = (P.moduleNameFromString "Module.F", [T.ideDtor "DtorA" "TypeA" Nothing
 moduleG = (P.moduleNameFromString "Module.G", [T.ideTypeClass "MyClass" P.kindType []])
 moduleH = (P.moduleNameFromString "Module.H", [T.ideValueOp "<$>" (P.Qualified P.ByNullSourcePos (Left "")) 0 Nothing Nothing])
 moduleI = (P.moduleNameFromString "Module.I", [T.ideTypeOp "~>" (P.Qualified P.ByNullSourcePos "") 0 Nothing Nothing])
+moduleDCtors = (P.moduleNameFromString "Module.WithDC", [T.ideType "Foo" Nothing [(P.ProperName "A", P.tyString), (P.ProperName "B", P.tyString)] ])
 
 modules :: ModuleMap [IdeDeclarationAnn]
 modules = Map.fromList [moduleA, moduleB]
+
+allModules :: ModuleMap [IdeDeclarationAnn]
+allModules = Map.fromList [moduleA, moduleB,moduleC,moduleD,moduleE,moduleF,moduleG,moduleH,moduleI,moduleDCtors]
 
 runEq :: Text -> [Module]
 runEq s = Map.toList (applyFilters [exactFilter s] modules)
@@ -40,6 +45,20 @@ runNamespace namespaces = Map.toList . applyFilters [namespaceFilter namespaces]
 
 runDeclaration :: [D.DeclarationType] -> [Module] -> [Module]
 runDeclaration decls = Map.toList . applyFilters [declarationTypeFilter (Set.fromList decls)] . Map.fromList
+
+runDependency :: [Text] -> [Module]
+runDependency = runDependency' "Whatever"
+
+runDependency' :: Text -> [Text] -> [Module]
+runDependency' currentModuleName imports = Map.toList $ applyFilters [dependencyFilter Nothing (P.ModuleName currentModuleName) (testParseImports currentModuleName imports)] allModules
+
+runDependencyQualified :: Text -> [Text] -> [Module]
+runDependencyQualified qualifier imports = Map.toList $ applyFilters [dependencyFilter (Just $ P.ModuleName qualifier) (P.ModuleName "Whatever") (testParseImports "Whatever" imports)] allModules
+
+testParseImports :: Text -> [Text] -> [Import]
+testParseImports currentModuleName imports = either (const []) (\(_, _, x, _) -> x) $ sliceImportSection moduleLines
+  where
+  moduleLines = "module " <> currentModuleName <> " where" : (imports <> [ "", "blah = 42" ])
 
 spec :: Spec
 spec = do
@@ -130,3 +149,45 @@ spec = do
     it "extracts modules by filtering `value`, and `valueoperator` declarations" $
       runDeclaration [D.Value, D.ValueOperator]
         [moduleA, moduleB, moduleD, moduleG, moduleE, moduleH] `shouldBe` [moduleA, moduleB, moduleH]
+  describe "dependencyFilter" $ do
+    describe "import types" $ do
+      it "filters by implicit imports" $ do
+        runDependency ["import Module.A", "import Module.C"] `shouldBe` [moduleA, moduleC]
+      it "filters by matching explicit value import" $ do
+        runDependency ["import Module.A (function1)"] `shouldBe` [moduleA]
+      it "filters by matching explicit value import from correct module" $ do
+        runDependency ["import Module.B (function1)"] `shouldBe` []
+      it "filters not matching explicit value import" $ do
+        runDependency ["import Module.A (function2)"] `shouldBe` []
+      it "filters out names in hiding import" $ do
+        runDependency ["import Module.A hiding (function1)"] `shouldBe` []
+      it "doesn't filter out not matching names in hiding import" $ do
+        runDependency ["import Module.A hiding (nonsense)"] `shouldBe` [moduleA]
+      it "filters by containing module" $ do
+        runDependency' "Module.A" ["import Module.Blah"] `shouldBe` [moduleA]
+    describe "declaration types" $ do
+      it "matches type" $ do
+        runDependency ["import Module.C (List)"] `shouldBe` [moduleC]
+      it "includes data constructor with (..)" $ do
+        runDependency ["import Module.F (TypeA(..))"] `shouldBe` [moduleF]
+      it "includes data constructor explicitly listed" $ do
+        runDependency ["import Module.F (TypeA(DtorA))"] `shouldBe` [moduleF]
+      it "does not include data constructor not explicitly listed" $ do
+        runDependency ["import Module.F (TypeA(BogusOtherConstructor))"] `shouldBe` []
+      it "does not include data constructor when only the type is imported" $ do
+        runDependency ["import Module.F (TypeA)"] `shouldBe` []
+      it "includes synonym" $ do
+        runDependency ["import Module.E (SFType)"] `shouldBe` [moduleE]
+      it "includes typeclass" $ do
+        runDependency ["import Module.G (class MyClass)"] `shouldBe` [moduleG]
+      it "includes value op" $ do
+        runDependency ["import Module.H ((<$>))"] `shouldBe` [moduleH]
+      it "includes type op" $ do
+        runDependency ["import Module.I (type (~>))"] `shouldBe` [moduleI]
+    describe "qualifiers" $ do
+      it "includes single qualified import and not unqualified things" $ do
+        runDependencyQualified "AA" [ "import Module.A as AA", "import Module.C"] `shouldBe` [moduleA]
+      it "includes multiple qualified imports" $ do
+        runDependencyQualified "Combined.Thing" [ "import Module.A as Combined.Thing", "import Module.C as Combined.Thing", "import Module.F"] `shouldBe` [moduleA, moduleC]
+      it "doesn't include qualified import when qualifier not specified" $ do
+        runDependency [ "import Module.AA as A"] `shouldBe` []
